# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: sourcing
    description: CABL sourcing and newsletter operations
  - name: admin
    description: CABL administration and catalog operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /quote-requests:
    post:
      operationId: createQuoteRequest
      tags: [sourcing]
      summary: Create a wholesale quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestInput"
      responses:
        "201":
          description: Quote request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /newsletter-subscriptions:
    post:
      operationId: subscribeNewsletter
      tags: [sourcing]
      summary: Subscribe an email address to CABL updates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscriptionInput"
      responses:
        "201":
          description: Newsletter subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscription"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/metadata:
    get:
      operationId: getAdminMetadata
      tags: [admin]
      summary: Get admin tables and relationships
      responses:
        "200":
          description: Admin metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMetadata"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
  /admin/tables/{table}/rows:
    get:
      operationId: listAdminRows
      tags: [admin]
      summary: List rows for an allowed admin table
      parameters:
        - $ref: "#/components/parameters/AdminTable"
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 25
        - name: offset
          in: query
          schema:
            type: number
            minimum: 0
            default: 0
      responses:
        "200":
          description: Table rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRows"
        "404":
          description: Unknown table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdminRow
      tags: [admin]
      summary: Create a row in an allowed admin table
      parameters:
        - $ref: "#/components/parameters/AdminTable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRowInput"
      responses:
        "201":
          description: Created row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRow"
        "400":
          description: Invalid row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/tables/{table}/rows/{id}:
    patch:
      operationId: updateAdminRow
      tags: [admin]
      summary: Update a row in an allowed admin table
      parameters:
        - $ref: "#/components/parameters/AdminTable"
        - $ref: "#/components/parameters/AdminRowId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRowInput"
      responses:
        "200":
          description: Updated row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRow"
        "400":
          description: Invalid row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Row not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAdminRow
      tags: [admin]
      summary: Delete a row from an allowed admin table
      parameters:
        - $ref: "#/components/parameters/AdminTable"
        - $ref: "#/components/parameters/AdminRowId"
      responses:
        "204":
          description: Row deleted
        "400":
          description: Delete blocked by a relationship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Row not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/seed:
    post:
      operationId: seedAdminData
      tags: [admin]
      summary: Seed CABL catalog and reference data
      responses:
        "200":
          description: Seed summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSeedResult"
        "500":
          description: Seed failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    AdminTable:
      name: table
      in: path
      required: true
      schema:
        type: string
    AdminRowId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    QuoteRequestItem:
      type: object
      properties:
        productId:
          type: number
          minimum: 1
        productName:
          type: string
          minLength: 1
        sku:
          type: ["string", "null"]
        quantity:
          type: number
          minimum: 1
      required:
        - productId
        - productName
        - sku
        - quantity
    QuoteRequestInput:
      type: object
      properties:
        customerName:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 5
          maxLength: 40
        businessName:
          type: ["string", "null"]
          maxLength: 160
        notes:
          type: ["string", "null"]
          maxLength: 1000
        items:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/QuoteRequestItem"
      required:
        - customerName
        - phone
        - businessName
        - notes
        - items
    QuoteRequest:
      type: object
      properties:
        id:
          type: number
        status:
          type: string
          enum: [received]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - createdAt
    NewsletterSubscriptionInput:
      type: object
      properties:
        email:
          type: string
          pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"
          maxLength: 254
      required:
        - email
    NewsletterSubscription:
      type: object
      properties:
        id:
          type: number
        status:
          type: string
          enum: [subscribed]
      required:
        - id
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AdminColumn:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, boolean, date, uuid, json, array]
        nullable:
          type: boolean
        primaryKey:
          type: boolean
        generated:
          type: boolean
        references:
          type: ["string", "null"]
      required: [key, label, type, nullable, primaryKey, generated, references]
    AdminRelation:
      type: object
      properties:
        column:
          type: string
        targetTable:
          type: string
        targetColumn:
          type: string
        kind:
          type: string
          enum: [belongs_to, has_many, many_to_many, parent]
        label:
          type: string
      required: [column, targetTable, targetColumn, kind, label]
    AdminTable:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        group:
          type: string
        primaryKey:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/AdminColumn"
        relations:
          type: array
          items:
            $ref: "#/components/schemas/AdminRelation"
      required: [key, label, group, primaryKey, columns, relations]
    AdminMetadata:
      type: object
      properties:
        tables:
          type: array
          items:
            $ref: "#/components/schemas/AdminTable"
      required: [tables]
    AdminRows:
      type: object
      properties:
        table:
          type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: number
      required: [table, rows, total]
    AdminRowInput:
      type: object
      properties:
        values:
          type: object
          additionalProperties: true
      required: [values]
    AdminRow:
      type: object
      properties:
        row:
          type: object
          additionalProperties: true
      required: [row]
    AdminTableCount:
      type: object
      properties:
        table:
          type: string
        label:
          type: string
        count:
          type: number
      required: [table, label, count]
    AdminDashboard:
      type: object
      properties:
        totalRows:
          type: number
        tableCounts:
          type: array
          items:
            $ref: "#/components/schemas/AdminTableCount"
        recentOrders:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [totalRows, tableCounts, recentOrders]
    AdminSeedResult:
      type: object
      properties:
        seeded:
          type: boolean
        inserted:
          type: number
        tables:
          type: object
          additionalProperties:
            type: number
      required: [seeded, inserted, tables]