# lib/api-zod/src/generated/types/adminColumn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminColumnType } from './adminColumnType';

export interface AdminColumn {
  key: string;
  label: string;
  type: AdminColumnType;
  nullable: boolean;
  primaryKey: boolean;
  generated: boolean;
  /** @nullable */
  references: string | null;
}