# lib/api-zod/src/generated/types/adminDashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminDashboardRecentOrdersItem } from './adminDashboardRecentOrdersItem';
import type { AdminTableCount } from './adminTableCount';

export interface AdminDashboard {
  totalRows: number;
  tableCounts: AdminTableCount[];
  recentOrders: AdminDashboardRecentOrdersItem[];
}