# lib/api-zod/src/generated/types/adminMetadata.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminTable } from './adminTable';

export interface AdminMetadata {
  tables: AdminTable[];
}