# lib/api-zod/src/generated/types/adminRelation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRelationKind } from './adminRelationKind';

export interface AdminRelation {
  column: string;
  targetTable: string;
  targetColumn: string;
  kind: AdminRelationKind;
  label: string;
}