# lib/api-zod/src/generated/types/adminRow.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRowRow } from './adminRowRow';

export interface AdminRow {
  row: AdminRowRow;
}