# lib/api-zod/src/generated/types/adminRows.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRowsRowsItem } from './adminRowsRowsItem';

export interface AdminRows {
  table: string;
  rows: AdminRowsRowsItem[];
  total: number;
}