# lib/api-zod/src/generated/types/adminSeedResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSeedResultTables } from './adminSeedResultTables';

export interface AdminSeedResult {
  seeded: boolean;
  inserted: number;
  tables: AdminSeedResultTables;
}