# lib/api-zod/src/generated/types/adminTable.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminColumn } from './adminColumn';
import type { AdminRelation } from './adminRelation';

export interface AdminTable {
  key: string;
  label: string;
  group: string;
  primaryKey: string;
  columns: AdminColumn[];
  relations: AdminRelation[];
}